IUT_iftar_app-main/IUT_iftar_app-main/requirements.txt
flask==2.3.3
requests==2.31.0
pyjwt==2.8.0
werkzeug==2.3.7
cachetools==5.3.1